+++ src/main/java/com/example/cataniaunited/game/board/GraphBuilder.java
@@ -1,7 +1,6 @@
 package com.example.cataniaunited.game.board;
 
 import com.example.cataniaunited.api.GameWebSocket;
-import com.example.cataniaunited.game.board.tileStuff.Tile;
 import org.jboss.logging.Logger;
 
 import java.util.ArrayList;
