+++ src/main/java/com/example/cataniaunited/game/board/SettlementPosition.java
@@ -1,5 +1,4 @@
 package com.example.cataniaunited.game.board;
-import com.example.cataniaunited.game.board.tileStuff.Tile;
 import com.example.cataniaunited.game.buildings.Building;
 
 import java.util.ArrayList;
